+++ core/src/oned/ODCode39Writer.cpp
@@ -64,7 +64,7 @@ Code39Writer::encode(const std::wstring& contents, int width, int height, BitMat
 	}
 
 	std::array<int, 9> widths = {};
-	int codeWidth = 24 + 1 + length;
+	size_t codeWidth = 24 + 1 + length;
 	for (size_t i = 0; i < length; ++i) {
 		int indexInString = IndexOf(ALPHABET_STRING, contents[i]);
 		if (indexInString < 0) {
